feat: filter title 변경 함수 추가
