+++ src/pages/like/index.tsx
@@ -8,12 +8,12 @@ import { idState, savedPostState } from "../../store/savePostStore";
 import Nav from "../../components/Nav/Nav";
 import Item from "../../components/common/Item/Item";
 import FilterBar from "../../components/Filter/filterBar";
-import FilterModal from "../../components/Filter";
+import FilterModal from "../../components/Filter/filterModal";
+import Filter from "../../components/Filter";
 
 export default function Like() {
   const savedPosts = useRecoilState(savedPostState);
   const postList = savedPosts[0];
-  const [isSearch, setIsSearch] = useState(false);
 
   return (
     <>
@@ -26,8 +26,7 @@ export default function Like() {
 
       <div className={style.main}>
         <Nav onclick={null} />
-        <FilterBar onclick={() => setIsSearch(!isSearch)} />
-        {isSearch && <FilterModal />}
+        <Filter />
 
         <div style={{ display: "flex", flexDirection: "column", gap: 10 }}>
           {postList.map((item, idx) => {
