+++ src/store/filterStore.ts
@@ -2,9 +2,8 @@ import { atom } from "recoil";
 import { filterText } from "../types/filter";
 
 export const filterTitleState = atom<filterText>({
-  key: "filterTitle",
+  key: "filterTitles",
   default: {
-    sortTitle: "최신순",
     dateTitle: "",
     tagTitle: "",
     contentTitle: "",
