+++ src/types/filter.ts
@@ -1,5 +1,4 @@
 export interface filterText {
-  sortTitle: string;
   dateTitle: string;
   tagTitle: string;
   contentTitle: string;
